WAYYY better lockscreen shell script
Now, it'll run as a daemon by default, monitoring for time spent idle.
At 9 minutes' idle time, it'll display a menu of options including
whether to lock now, or turn off the lockscreen for ten minutes/an hour/
a day/until I log out.

It can also lock immediately with the "suspend" argument.